Добавлены группы пользователей, полностью исправлен счетчик непрочитанных сообщений для групп. Добавить функционал ролям пользователей в группах, назначение ролей

+++ MessageStorageService/src/main/java/com/alexgls/springboot/messagestorageservice/dto/ReadMessagePayload.java
@@ -1,7 +1,7 @@
 package com.alexgls.springboot.messagestorageservice.dto;
 
 public record ReadMessagePayload(
-        int messageId,
+        long messageId,
         int senderId,
         int chatId
 ) {
